Move ArticleMain utils to common utils
