+++ src/app/lib/utilities/parseAssetData/index.js
